+++ blocks/contribute/exemples.php
@@ -1,5 +1,17 @@
 <?php
 return [
   "Standard" => [
-  ]
+		'background_color' => '#f8f5ef',
+		'title' => 'Contribuez dès maintenant',
+		'buttons' => [
+			'display' => 'buttons',
+			'items' => [
+				[
+					'href' => '#',
+					'text' => __('Trouver un moyen de participer', 'telabotanica'),
+					'modifiers' => 'orange'
+				]
+			]
+		]
+	]
 ];