+++ functions.php
@@ -111,6 +111,7 @@ function telabotanica_setup() {
 	 */
 	add_theme_support( 'post-thumbnails' );
 	set_post_thumbnail_size( 220, 160, array( 'center', 'center') );
+	add_image_size( 'medium_square', 250, 250, array( 'center', 'center') );
 	add_image_size( 'home-latest-post', 600, 365, array( 'center', 'center') );
 	add_image_size( 'home-post-thumbnail', 65, 50, array( 'center', 'center') );